+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the "midcentury" extension will be documented in this fil
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [Unreleased]
+
+- Add **Midcentury Light**, a light companion theme derived from the same palette
+
 ## [1.0.0] - 2023-07-27
 
 - Initial release